+++ tests/Component/FileComponent/Php/Property/PropertyTest.php
@@ -18,6 +18,9 @@ class PropertyTest extends TestCase
 {
     /**
      * @covers ::getVisibility
+     * @covers ::setVisibility
+     * @covers ::setIsStatic
+     * @covers ::isStatic
      * @covers ::getContent
      * @covers ::__construct
      *
@@ -31,7 +34,7 @@ class PropertyTest extends TestCase
         $description = 'Bar Baz.';
         $valueContent = '\'foo\'';
         $value = $this->createMock(ValueInterface::class);
-        $value->expects(static::once())
+        $value->expects(static::exactly(2))
             ->method('getContent')
             ->willReturn($valueContent);
         $subject = new Property($name, $visibility, $type, $description, $value);
@@ -44,6 +47,22 @@ class PropertyTest extends TestCase
             ' * @var mixed' . PHP_EOL .
             ' */' . PHP_EOL .
             $visibility . ' $' . $name . ' = ' . $valueContent . ';' . PHP_EOL .
+            PHP_EOL,
+            $subject->getContent()
+        );
+
+        $this->assertEquals(false, $subject->isStatic());
+        $subject->setIsStatic(true);
+        $this->assertEquals(true, $subject->isStatic());
+        $subject->setVisibility('public');
+
+        $this->assertEquals(
+            '/**' . PHP_EOL .
+            ' * ' . $description . PHP_EOL .
+            ' * ' . PHP_EOL .
+            ' * @var mixed' . PHP_EOL .
+            ' */' . PHP_EOL .
+            'public static $' . $name . ' = ' . $valueContent . ';' . PHP_EOL .
             PHP_EOL,
             $subject->getContent()
         );