+++ tests/Component/FileComponent/Php/Reference/UseReferenceTest.php
@@ -17,6 +17,7 @@ class UseReferenceTest extends TestCase
 {
     /**
      * @covers ::getName
+     * @covers ::setName
      * @covers ::getContent
      * @covers ::__construct
      *
@@ -30,6 +31,14 @@ class UseReferenceTest extends TestCase
         $this->assertEquals($name, $subject->getName());
         $this->assertEquals(
             'use ' . $name . ';' . PHP_EOL,
+            $subject->getContent()
+        );
+
+        $subject->setName('foo');
+        $this->assertEquals('foo', $subject->getName());
+
+        $this->assertEquals(
+            'use foo;' . PHP_EOL,
             $subject->getContent()
         );
     }