+++ CHANGELOG.md
@@ -4,10 +4,45 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2020-10-17
+
+### Added
+#### Files
+- `setName` method for files.
+
+#### Variables
+- `setName` method for variables.
+- `setDescription` method for variables.
+- `setType` method for variables.
+- `setValue` method for variables.
+- `setValue` method for values.
+
+#### References
+- `setName` method for references.
+
+#### Properties
+- `setVisibility` method for properties.
+- `setIsStatic` method for properties.
+- `isStatic` method for properties.
+
+#### Methods
+- `setIsFinal` method for methods.
+- `isFinal` method for methods.
+
+#### Definitions
+- `setName` method for definitions.
+- `setNamespace` method for definitions.
+- `setExtends` method for definitions.
+
+#### Classes
+- `setIsFinal` method for classes.
+- `isFinal` method for classes.
+
 ## 1.0.0 - 2020-10-17
 
 ### Added
 - The initial implementation of the package.
 
 # Versions
-- [1.0.0 > Unreleased](https://github.com/grizz-it/ast/compare/1.0.0...HEAD)
+- [1.1.0 > Unreleased](https://github.com/grizz-it/ast/compare/1.1.0...HEAD)
+- [1.0.0 > 2.0.0](https://github.com/grizz-it/ast/compare/1.0.0...1.1.0)